+++ ios/AppLovinMAX.m
@@ -84,6 +84,16 @@
 static NSString *const SDK_TAG = @"AppLovinSdk";
 static NSString *const TAG = @"AppLovinMAX";
 
+static NSString *const USER_GEOGRAPHY_GDPR = @"G";
+static NSString *const USER_GEOGRAPHY_OTHER = @"O";
+static NSString *const USER_GEOGRAPHY_UNKNOWN = @"U";
+
+static NSString *const APP_TRACKING_STATUS_NOTDETERMINED = @"N";
+static NSString *const APP_TRACKING_STATUS_RESTRICTED = @"R";
+static NSString *const APP_TRACKING_STATUS_DENIED = @"D";
+static NSString *const APP_TRACKING_STATUS_AUTHORIZED = @"A";
+static NSString *const APP_TRACKING_STATUS_UNAVAILABLE = @"U";
+
 static NSString *const ON_BANNER_AD_LOADED_EVENT = @"OnBannerAdLoadedEvent";
 static NSString *const ON_BANNER_AD_LOAD_FAILED_EVENT = @"OnBannerAdLoadFailedEvent";
 static NSString *const ON_BANNER_AD_CLICKED_EVENT = @"OnBannerAdClickedEvent";
@@ -216,29 +226,53 @@ RCT_EXPORT_METHOD(initialize:(NSString *)pluginVersion :(NSString *)sdkKey :(RCT
     ALSdkSettings *settings = [[ALSdkSettings alloc] init];
 
     // Deprecated consent flow which automatically moves to the new flow
-    if ( self.consentFlowEnabledToSet.boolValue )
+    if ( self.consentFlowEnabledToSet )
     {
-        settings.consentFlowSettings.enabled = YES;
-        settings.consentFlowSettings.privacyPolicyURL = self.privacyPolicyURLToSet;
-        settings.consentFlowSettings.termsOfServiceURL = self.termsOfServiceURLToSet;
-
+        settings.consentFlowSettings.enabled = self.consentFlowEnabledToSet.boolValue;
         self.consentFlowEnabledToSet = nil;
-        self.privacyPolicyURLToSet = nil;
-        self.termsOfServiceURLToSet = nil;
+    }
+
+    if ( settings.consentFlowSettings.enabled )
+    {
+        if ( self.privacyPolicyURLToSet )
+        {
+            settings.consentFlowSettings.privacyPolicyURL = self.privacyPolicyURLToSet;
+            self.privacyPolicyURLToSet = nil;
+        }
+
+        if (self.termsOfServiceURLToSet )
+        {
+            settings.consentFlowSettings.termsOfServiceURL = self.termsOfServiceURLToSet;
+            self.termsOfServiceURLToSet = nil;
+        }
     }
 
     // New terms and privacy policy flow
-    if ( self.termsAndPrivacyPolicyFlowEnabledToSet.boolValue )
+    if ( self.termsAndPrivacyPolicyFlowEnabledToSet )
     {
-        settings.termsAndPrivacyPolicyFlowSettings.enabled = YES;
-        settings.termsAndPrivacyPolicyFlowSettings.privacyPolicyURL = self.privacyPolicyURLToSet;
-        settings.termsAndPrivacyPolicyFlowSettings.termsOfServiceURL = self.termsOfServiceURLToSet;
-        settings.termsAndPrivacyPolicyFlowSettings.debugUserGeography = [self toAppLovinConsentFlowUserGeography: self.debugUserGeographyToSet];
-
+        settings.termsAndPrivacyPolicyFlowSettings.enabled = self.termsAndPrivacyPolicyFlowEnabledToSet.boolValue;
         self.termsAndPrivacyPolicyFlowEnabledToSet = nil;
-        self.privacyPolicyURLToSet = nil;
-        self.termsOfServiceURLToSet = nil;
-        self.debugUserGeographyToSet = nil;
+    }
+
+    if ( settings.termsAndPrivacyPolicyFlowSettings.enabled )
+    {
+        if ( self.privacyPolicyURLToSet )
+        {
+            settings.termsAndPrivacyPolicyFlowSettings.privacyPolicyURL = self.privacyPolicyURLToSet;
+            self.privacyPolicyURLToSet = nil;
+        }
+
+        if ( self.termsOfServiceURLToSet )
+        {
+            settings.termsAndPrivacyPolicyFlowSettings.termsOfServiceURL = self.termsOfServiceURLToSet;
+            self.termsOfServiceURLToSet = nil;
+        }
+
+        if ( self.debugUserGeographyToSet )
+        {
+            settings.termsAndPrivacyPolicyFlowSettings.debugUserGeography = [self toAppLovinConsentFlowUserGeography: self.debugUserGeographyToSet];
+            self.debugUserGeographyToSet = nil;
+        }
     }
 
     // Initialize SDK
@@ -2011,11 +2045,11 @@ RCT_EXPORT_METHOD(setAppOpenAdLocalExtraParameter:(NSString *)adUnitIdentifier :
 
 - (ALConsentFlowUserGeography)toAppLovinConsentFlowUserGeography:(NSString *)userGeography
 {
-    if ( [@"G" al_isEqualToStringIgnoringCase: userGeography] )
+    if ( [USER_GEOGRAPHY_GDPR al_isEqualToStringIgnoringCase: userGeography] )
     {
         return ALConsentFlowUserGeographyGDPR;
     }
-    else if ( [@"O" al_isEqualToStringIgnoringCase: userGeography] )
+    else if ( [USER_GEOGRAPHY_OTHER al_isEqualToStringIgnoringCase: userGeography] )
     {
         return ALConsentFlowUserGeographyOther;
     }
@@ -2027,36 +2061,36 @@ RCT_EXPORT_METHOD(setAppOpenAdLocalExtraParameter:(NSString *)adUnitIdentifier :
 {
     if ( ALConsentFlowUserGeographyGDPR == userGeography )
     {
-        return @"G";
+        return USER_GEOGRAPHY_GDPR;
     }
     else if ( ALConsentFlowUserGeographyOther == userGeography )
     {
-        return @"O";
+        return USER_GEOGRAPHY_OTHER;
     }
 
-    return @"U";
+    return USER_GEOGRAPHY_UNKNOWN;
 }
 
 - (NSString *)fromAppLovinAppTrackingStatus:(ALAppTrackingTransparencyStatus)status
 {
     if ( ALAppTrackingTransparencyStatusNotDetermined == status )
     {
-        return @"N";
+        return APP_TRACKING_STATUS_NOTDETERMINED;
     }
     else if ( ALAppTrackingTransparencyStatusRestricted == status )
     {
-        return @"R";
+        return APP_TRACKING_STATUS_RESTRICTED;
     }
     else if ( ALAppTrackingTransparencyStatusDenied == status )
     {
-        return @"D";
+        return APP_TRACKING_STATUS_DENIED;
     }
     else if ( ALAppTrackingTransparencyStatusAuthorized == status )
     {
-        return @"A";
+        return APP_TRACKING_STATUS_AUTHORIZED;
     }
 
-    return @"U"; // ALAppTrackingTransparencyStatusUnavailable
+    return APP_TRACKING_STATUS_UNAVAILABLE;
 }
 
 #pragma mark - Ad Info